+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Changed
 ### Fixed
 
-- (@polvalente) Fix issue where Base.decode64 :error return could bleed trough as {:ok, :error} (#237)
+- (@polvalente) Fix issue where Base.decode64 made peek_claims and peek_header return out of spec (#237)
 
 ## [2.1.0] - 2019-05-27